phase 4 - battle modes and splitting up logic to different paths
