+++ fp/battle/protocol.py
@@ -1,18 +1,12 @@
 import re
 import json
-from copy import deepcopy, copy
+from copy import deepcopy
 import logging
 
 from fp import constants
-from fp.constants import BattleType
 from fp.data import all_move_json
 from fp.data import pokedex
-from fp.data.pkmn_sets import (
-    SmogonSets,
-    RandomBattleTeamDatasets,
-    TeamDatasets,
-)
-from fp.battle.state import Pokemon, Battler, Battle
+from fp.battle.state import Pokemon, Battle
 from fp.battle.state import LastUsedMove
 from fp.search.poke_engine_helpers import poke_engine_get_damage_rolls
 from fp.battle.helpers import (
@@ -56,20 +50,6 @@ SIDE_CONDITION_DEFAULT_DURATION = {
     constants.MIST: 5,
     constants.TAILWIND: 4,
 }
-
-
-def crit_rate_for_generation(generation):
-    if generation == "gen1":
-        return 205 / 105
-    elif generation in [
-        "gen2",
-        "gen3",
-        "gen4",
-        "gen5",
-    ]:
-        return 2.0
-    else:
-        return 1.5
 
 
 def remove_volatile(pkmn, volatile):
@@ -352,16 +332,7 @@ def switch_or_drag(battle, split_msg, switch_or_drag="switch"):
     if pkmn is None:
         pkmn = Pokemon.from_switch_string(split_msg[3], nickname=nickname)
 
-        # for standard battles gen4 and lower
-        # we want to add the new pokemon to the datasets as they are revealed
-        # because there is no teampreview
-        if (
-            battle.battle_type == BattleType.STANDARD_BATTLE
-            and not battle.gen.has_team_preview
-        ):
-            SmogonSets.add_new_pokemon(pkmn.name)
-            TeamDatasets.add_new_pokemon(pkmn.name)
-            logger.info("Adding new pokemon '{}' to the datasets".format(pkmn.name))
+        battle.mode.add_revealed_pokemon(battle, pkmn)
 
         # some pokemon do not reveal their forme during team preview. Arceus, Silvally, Genesect, etc.
         # if this is the case, they would have been given a flag during team preview, and we can pull them out here
@@ -663,82 +634,9 @@ def move(battle, split_msg):
         "zoroark"
     ) or side.find_pokemon_in_reserves("zoroarkhisui")
 
-    # in battle factory we can deduce that there is a zoroark in front of us
-    # if we see a move that is not in the known moveset and a zoroark is in the reserves
-    if (
-        is_opponent(battle, split_msg)
-        and zoroark_from_reserves is not None
-        and "transform" not in pkmn.volatile_statuses
-        and battle.battle_type
-        in [BattleType.BATTLE_FACTORY, BattleType.STANDARD_BATTLE]
-        and move_name not in TeamDatasets.get_all_possible_moves(pkmn)
-        and move_name in TeamDatasets.get_all_possible_moves(zoroark_from_reserves)
-        and "from" not in split_msg[-1]
-    ):
-        logger.info(
-            "{} using {} means it is {}".format(
-                pkmn.name, move_name, zoroark_from_reserves.name
-            )
-        )
-        _switch_active_with_zoroark_from_reserves(side, zoroark_from_reserves)
-
-        # the rest of this function uses `pkmn`, so we need to set it to the correct pkmn
-        pkmn = zoroark_from_reserves
-
-    # in randombattles we can deduce that there is a zoroark in front of us
-    # if we see a move that is not in the known moveset, even if there is no
-    # zoroark is in the reserves
-    if (
-        is_opponent(battle, split_msg)
-        and battle.battle_type == BattleType.RANDOM_BATTLE
-        and "transform" not in pkmn.volatile_statuses
-        and move_name not in RandomBattleTeamDatasets.get_all_possible_moves(pkmn)
-        and "from" not in split_msg[-1]
-    ):
-        actual_zoroark = None
-        zoroark_hisui = Pokemon("zoroarkhisui", 100)
-        zoroark_regular = Pokemon("zoroark", 100)
-        if (
-            zoroark_from_reserves is not None
-            and move_name
-            in RandomBattleTeamDatasets.get_all_possible_moves(zoroark_from_reserves)
-        ):
-            actual_zoroark = zoroark_from_reserves
-
-        elif (
-            battle.gen.has_team_preview
-            and zoroark_from_reserves is None
-            and move_name
-            in RandomBattleTeamDatasets.get_all_possible_moves(zoroark_hisui)
-        ):
-            actual_zoroark = zoroark_hisui
-            actual_zoroark.level = RandomBattleTeamDatasets.predict_set(
-                actual_zoroark
-            ).pkmn_set.level
-            side.reserve.append(actual_zoroark)
-
-        elif (
-            battle.gen.has_team_preview
-            and zoroark_from_reserves is None
-            and move_name
-            in RandomBattleTeamDatasets.get_all_possible_moves(zoroark_regular)
-        ):
-            actual_zoroark = zoroark_regular
-            actual_zoroark.level = RandomBattleTeamDatasets.predict_set(
-                actual_zoroark
-            ).pkmn_set.level
-            side.reserve.append(actual_zoroark)
-
-        if actual_zoroark is not None:
-            logger.info(
-                "{} using {} means it is {}".format(
-                    pkmn.name, move_name, actual_zoroark.name
-                )
-            )
-            _switch_active_with_zoroark_from_reserves(side, actual_zoroark)
-
-            # the rest of this function uses `pkmn`, so we need to set it to the correct pkmn
-            pkmn = actual_zoroark
+    pkmn = battle.mode.check_zoroark_from_move(
+        battle, side, pkmn, move_name, split_msg, zoroark_from_reserves
+    )
 
     if (
         any(msg == "[from]Sleep Talk" for msg in split_msg)
@@ -1720,136 +1618,7 @@ def immune(battle, split_msg):
             == 0
         )
     ):
-        # Battle Factory: Zoroark must be in the reserves
-        # and must be immune to the last used move by the bot
-        if (
-            battle.battle_type == BattleType.BATTLE_FACTORY
-            and zoroark_from_reserves is not None
-            and type_effectiveness_modifier(
-                all_move_json[battle.user.last_used_move.move][constants.TYPE],
-                zoroark_from_reserves.types,
-            )
-            == 0
-        ):
-            logger.info(
-                "{} was immune to {} when it shouldn't be - it is {}".format(
-                    pkmn.name,
-                    battle.user.last_used_move.move,
-                    zoroark_from_reserves.name,
-                )
-            )
-            _switch_active_with_zoroark_from_reserves(side, zoroark_from_reserves)
-
-        # Random Battle: Zoroark may be in the reserves so we need to check the move type
-        # that it was immune to
-        elif battle.battle_type == BattleType.RANDOM_BATTLE:
-            actual_zoroark = None
-            zoroark_hisui = Pokemon("zoroarkhisui", 100)
-            zoroark_regular = Pokemon("zoroark", 100)
-
-            # zoroark was in the reserves - just use that one
-            if (
-                zoroark_from_reserves is not None
-                and type_effectiveness_modifier(
-                    all_move_json[battle.user.last_used_move.move][constants.TYPE],
-                    zoroark_from_reserves.types,
-                )
-                == 0
-            ):
-                actual_zoroark = zoroark_from_reserves
-
-            # hisui zoroark
-            elif (
-                zoroark_from_reserves is None
-                and type_effectiveness_modifier(
-                    all_move_json[battle.user.last_used_move.move][constants.TYPE],
-                    zoroark_hisui.types,
-                )
-                == 0
-                and zoroark_hisui.name in RandomBattleTeamDatasets.pkmn_sets
-            ):
-                actual_zoroark = zoroark_hisui
-                actual_zoroark.level = RandomBattleTeamDatasets.predict_set(
-                    actual_zoroark
-                ).pkmn_set.level
-                side.reserve.append(actual_zoroark)
-
-            # regular zoroark
-            elif (
-                zoroark_from_reserves is None
-                and type_effectiveness_modifier(
-                    all_move_json[battle.user.last_used_move.move][constants.TYPE],
-                    zoroark_regular.types,
-                )
-                == 0
-                and zoroark_regular.name in RandomBattleTeamDatasets.pkmn_sets
-            ):
-                actual_zoroark = zoroark_regular
-                actual_zoroark.level = RandomBattleTeamDatasets.predict_set(
-                    actual_zoroark
-                ).pkmn_set.level
-                side.reserve.append(actual_zoroark)
-
-            # if we found a zoroark from one of those branches
-            if actual_zoroark is not None:
-                logger.info(
-                    "{} was immune to {} when it shouldn't be - it is {}".format(
-                        pkmn.name,
-                        battle.user.last_used_move.move,
-                        actual_zoroark.name,
-                    )
-                )
-                _switch_active_with_zoroark_from_reserves(side, actual_zoroark)
-
-
-def _switch_active_with_zoroark_from_reserves(
-    opponent_side: Battler, zoroark_from_reserves: Pokemon
-):
-    """
-    This is called when we are 100% sure that the opponent's active pkmn is a zoroark
-    This swaps the active pkmn with the zoroark from the reserves
-
-    Assumptions:
-        - The `zoroark_from_reserves` MUST be in `opponent_side.reserve`
-    """
-    pkmn = opponent_side.active
-
-    # any moves used by this pkmn since switching in need to be removed because we cannot guarantee that they
-    # belong to this pkmn
-    for mv in pkmn.moves_used_since_switch_in:
-        logger.info(
-            "Removing {} from {}'s moves because it is {}".format(
-                mv, pkmn.name, zoroark_from_reserves.name
-            )
-        )
-        pkmn.remove_move(mv)
-        if zoroark_from_reserves.get_move(mv) is None:
-            zoroark_from_reserves.add_move(mv)
-
-    # set attributes on zoroark that were on the pokemon that we thought was zoroark
-    # and clear those attributes from the pokemon that we thought was zoroark
-    pkmn_hp_percent = float(pkmn.hp) / pkmn.max_hp
-    zoroark_from_reserves.hp = zoroark_from_reserves.max_hp * pkmn_hp_percent
-    zoroark_from_reserves.boosts = copy(pkmn.boosts)
-    zoroark_from_reserves.status = pkmn.status
-    zoroark_from_reserves.volatile_statuses = copy(pkmn.volatile_statuses)
-    zoroark_from_reserves.terastallized = pkmn.terastallized
-    zoroark_from_reserves.tera_type = pkmn.tera_type
-    pkmn.boosts.clear()
-    pkmn.status = None
-    pkmn.volatile_statuses.clear()
-    pkmn.volatile_status_durations.clear()
-
-    if pkmn.terastallized:
-        pkmn.terastallized = False
-        pkmn.tera_type = None
-
-    zoroark_from_reserves.zoroark_disguised_as = pkmn.name
-
-    # swap the pkmn places
-    opponent_side.reserve.append(pkmn)
-    opponent_side.active = zoroark_from_reserves
-    opponent_side.reserve.remove(zoroark_from_reserves)
+        battle.mode.check_zoroark_from_immune(battle, side, pkmn, zoroark_from_reserves)
 
 
 def update_ability(battle, split_msg):
