+++ fp/run_battle.py
@@ -1,66 +1,13 @@
-import json
-import asyncio
-import concurrent.futures
-from copy import deepcopy
 import logging
 
-from fp.data.pkmn_sets import RandomBattleTeamDatasets, TeamDatasets
-from fp.data.pkmn_sets import SmogonSets
 from fp import constants
-from fp.constants import BattleType
 from fp.config import FoulPlayConfig, SaveReplay
-from fp.battle.state import LastUsedMove, Pokemon, Battle
-from fp.battle.protocol import async_update_battle, process_battle_updates
-from fp.battle.helpers import normalize_name
+from fp.battle.protocol import async_update_battle
 from fp.format_spec import FormatSpec
-from fp.search.main import find_best_move
-
-from fp.websocket_client import PSWebsocketClient
+from fp.modes import battle_mode
+from fp.modes.base import async_pick_move
 
 logger = logging.getLogger(__name__)
-
-
-def format_decision(battle, decision):
-    # Formats a decision for communication with Pokemon-Showdown
-    # If the move can be used as a Z-Move, it will be
-
-    if decision.startswith(constants.SWITCH_STRING + " "):
-        switch_pokemon = decision.split("switch ")[-1]
-        for pkmn in battle.user.reserve:
-            if pkmn.name == switch_pokemon:
-                message = "/switch {}".format(pkmn.index)
-                break
-        else:
-            raise ValueError("Tried to switch to: {}".format(switch_pokemon))
-    else:
-        tera = False
-        mega = False
-        if decision.endswith("-tera"):
-            decision = decision.replace("-tera", "")
-            tera = True
-        elif decision.endswith("-mega"):
-            decision = decision.replace("-mega", "")
-            mega = True
-        message = "/choose move {}".format(decision)
-
-        if battle.user.active.can_mega_evo and mega:
-            message = "{} {}".format(message, constants.MEGA)
-        elif battle.user.active.can_ultra_burst:
-            message = "{} {}".format(message, constants.ULTRA_BURST)
-
-        # only dynamax on last pokemon
-        if battle.user.active.can_dynamax and all(
-            p.hp == 0 for p in battle.user.reserve
-        ):
-            message = "{} {}".format(message, constants.DYNAMAX)
-
-        if tera:
-            message = "{} {}".format(message, constants.TERASTALLIZE)
-
-        if battle.user.active.get_move(decision).can_z:
-            message = "{} {}".format(message, constants.ZMOVE)
-
-    return [message, str(battle.rqid)]
 
 
 def battle_is_finished(battle_tag, msg):
@@ -71,240 +18,11 @@ def battle_is_finished(battle_tag, msg):
     )
 
 
-def extract_battle_factory_tier_from_msg(msg):
-    start = msg.find("Battle Factory Tier: ") + len("Battle Factory Tier: ")
-    end = msg.find("</b>", start)
-    tier_name = msg[start:end]
-
-    return normalize_name(tier_name)
-
-
-async def async_pick_move(battle):
-    battle_copy = deepcopy(battle)
-    if not battle_copy.team_preview:
-        battle_copy.user.update_from_request_json(battle_copy.request_json)
-
-    loop = asyncio.get_event_loop()
-    with concurrent.futures.ThreadPoolExecutor() as pool:
-        best_move = await loop.run_in_executor(pool, find_best_move, battle_copy)
-    battle.user.last_selected_move = LastUsedMove(
-        battle.user.active.name,
-        best_move.removesuffix("-tera").removesuffix("-mega"),
-        battle.turn,
-    )
-    return format_decision(battle_copy, best_move)
-
-
-async def handle_team_preview(battle, ps_websocket_client):
-    battle_copy = deepcopy(battle)
-    battle_copy.user.active = Pokemon.get_dummy()
-    battle_copy.opponent.active = Pokemon.get_dummy()
-    battle_copy.team_preview = True
-
-    best_move = await async_pick_move(battle_copy)
-
-    # because we copied the battle before sending it in, we need to update the last selected move here
-    pkmn_name = battle.user.reserve[int(best_move[0].split()[1]) - 1].name
-    battle.user.last_selected_move = LastUsedMove(
-        "teampreview", "switch {}".format(pkmn_name), battle.turn
-    )
-
-    size_of_team = len(battle.user.reserve) + 1
-    team_list_indexes = list(range(1, size_of_team))
-    choice_digit = int(best_move[0].split()[-1])
-
-    team_list_indexes.remove(choice_digit)
-    message = [
-        "/team {}{}|{}".format(
-            choice_digit, "".join(str(x) for x in team_list_indexes), battle.rqid
-        )
-    ]
-
-    await ps_websocket_client.send_message(battle.battle_tag, message)
-
-
-async def get_battle_tag_and_opponent(ps_websocket_client: PSWebsocketClient):
-    while True:
-        msg = await ps_websocket_client.receive_message()
-        split_msg = msg.split("|")
-        first_msg = split_msg[0]
-        if "battle" in first_msg:
-            battle_tag = first_msg.replace(">", "").strip()
-            user_name = FoulPlayConfig.username
-            opponent_name = (
-                split_msg[4].replace(user_name, "").replace("vs.", "").strip()
-            )
-            logger.info("Initialized {} against: {}".format(battle_tag, opponent_name))
-            return battle_tag, opponent_name
-
-
-async def start_battle_common(
-    ps_websocket_client: PSWebsocketClient, pokemon_battle_type
-):
-    battle_tag, opponent_name = await get_battle_tag_and_opponent(ps_websocket_client)
-    if FoulPlayConfig.log_to_file:
-        FoulPlayConfig.file_log_handler.do_rollover(
-            "{}_{}.log".format(battle_tag, opponent_name)
-        )
-
-    battle = Battle(battle_tag)
-    battle.opponent.account_name = opponent_name
-    battle.pokemon_format = pokemon_battle_type
-    battle.generation = battle.format_spec.generation
-    battle.battle_type = battle.format_spec.battle_type
-
-    # wait until the opponent's identifier is received. This will be `p1` or `p2`.
-    #
-    # e.g.
-    # '>battle-gen9randombattle-44733
-    # |player|p1|OpponentName|2|'
-    while True:
-        msg = await ps_websocket_client.receive_message()
-        if "|player|" in msg and battle.opponent.account_name in msg:
-            battle.opponent.name = msg.split("|")[2]
-            battle.user.name = constants.ID_LOOKUP[battle.opponent.name]
-            break
-
-    return battle, msg
-
-
-async def get_first_request_json(
-    ps_websocket_client: PSWebsocketClient, battle: Battle
-):
-    while True:
-        msg = await ps_websocket_client.receive_message()
-        msg_split = msg.split("|")
-        if msg_split[1].strip() == "request" and msg_split[2].strip():
-            user_json = json.loads(msg_split[2].strip("'"))
-            battle.request_json = user_json
-            battle.user.initialize_first_turn_user_from_json(user_json)
-            battle.rqid = user_json[constants.RQID]
-            return
-
-
-async def start_random_battle(
-    ps_websocket_client: PSWebsocketClient, pokemon_battle_type
-):
-    battle, msg = await start_battle_common(ps_websocket_client, pokemon_battle_type)
-    RandomBattleTeamDatasets.initialize(pokemon_battle_type)
-
-    while True:
-        if constants.START_STRING in msg:
-            battle.started = True
-
-            # hold onto some messages to apply after we get the request JSON
-            # omit the bot's switch-in message because we won't need that
-            # parsing the request JSON will set the bot's active pkmn
-            battle.msg_list = [
-                m
-                for m in msg.split(constants.START_STRING)[1].strip().split("\n")
-                if not (m.startswith("|switch|{}".format(battle.user.name)))
-            ]
-            break
-        msg = await ps_websocket_client.receive_message()
-
-    await get_first_request_json(ps_websocket_client, battle)
-
-    # apply the messages that were held onto
-    process_battle_updates(battle)
-
-    best_move = await async_pick_move(battle)
-    await ps_websocket_client.send_message(battle.battle_tag, best_move)
-
-    return battle
-
-
-async def start_standard_battle(
-    ps_websocket_client: PSWebsocketClient, pokemon_battle_type, team_dict
-):
-    battle, msg = await start_battle_common(ps_websocket_client, pokemon_battle_type)
-    battle.user.team_dict = team_dict
-
-    if not battle.gen.has_team_preview:
-        while True:
-            if constants.START_STRING in msg:
-                battle.started = True
-
-                # hold onto some messages to apply after we get the request JSON
-                # omit the bot's switch-in message because we won't need that
-                # parsing the request JSON will set the bot's active pkmn
-                battle.msg_list = [
-                    m
-                    for m in msg.split(constants.START_STRING)[1].strip().split("\n")
-                    if not (m.startswith("|switch|{}".format(battle.user.name)))
-                ]
-                break
-            msg = await ps_websocket_client.receive_message()
-
-        await get_first_request_json(ps_websocket_client, battle)
-
-        unique_pkmn_names = set(
-            [p.name for p in battle.user.reserve] + [battle.user.active.name]
-        )
-        SmogonSets.initialize(
-            FoulPlayConfig.smogon_stats or pokemon_battle_type, unique_pkmn_names
-        )
-        TeamDatasets.initialize(pokemon_battle_type, unique_pkmn_names)
-
-        # apply the messages that were held onto
-        process_battle_updates(battle)
-
-        best_move = await async_pick_move(battle)
-        await ps_websocket_client.send_message(battle.battle_tag, best_move)
-
-    else:
-        while constants.START_TEAM_PREVIEW not in msg:
-            msg = await ps_websocket_client.receive_message()
-
-        preview_string_lines = msg.split(constants.START_TEAM_PREVIEW)[-1].split("\n")
-
-        opponent_pokemon = []
-        for line in preview_string_lines:
-            if not line:
-                continue
-
-            split_line = line.split("|")
-            if (
-                split_line[1] == constants.TEAM_PREVIEW_POKE
-                and split_line[2].strip() == battle.opponent.name
-            ):
-                opponent_pokemon.append(split_line[3])
-
-        await get_first_request_json(ps_websocket_client, battle)
-        battle.initialize_team_preview(opponent_pokemon, pokemon_battle_type)
-        battle.during_team_preview()
-
-        unique_pkmn_names = set(
-            p.name for p in battle.opponent.reserve + battle.user.reserve
-        )
-
-        if battle.battle_type == BattleType.BATTLE_FACTORY:
-            tier_name = extract_battle_factory_tier_from_msg(msg)
-            logger.info("Battle Factory Tier: {}".format(tier_name))
-            TeamDatasets.initialize(
-                pokemon_battle_type,
-                unique_pkmn_names,
-                battle_factory_tier_name=tier_name,
-            )
-        else:
-            SmogonSets.initialize(
-                FoulPlayConfig.smogon_stats or pokemon_battle_type, unique_pkmn_names
-            )
-            TeamDatasets.initialize(pokemon_battle_type, unique_pkmn_names)
-
-        await handle_team_preview(battle, ps_websocket_client)
-
-    return battle
-
-
 async def start_battle(ps_websocket_client, pokemon_battle_type, team_dict):
     format_spec = FormatSpec.from_format_string(pokemon_battle_type)
-    if format_spec.battle_type == BattleType.RANDOM_BATTLE:
-        battle = await start_random_battle(ps_websocket_client, pokemon_battle_type)
-    else:
-        battle = await start_standard_battle(
-            ps_websocket_client, pokemon_battle_type, team_dict
-        )
+    battle = await battle_mode(format_spec.battle_type).start_battle(
+        ps_websocket_client, pokemon_battle_type, team_dict
+    )
 
     await ps_websocket_client.send_message(battle.battle_tag, ["hf"])
     await ps_websocket_client.send_message(battle.battle_tag, ["/timer on"])
